+++ next-docs/public/examples/search/Default.tsx
@@ -1,9 +1,8 @@
-import CommandPalette from './components/CommandPalette';
+import Search from './components/Search';
 import { filterItems, getItemIndex } from "./utils/utils";
 import { useState } from "react";
 
 const Example = () => {
-    const [page, setPage] = useState<"root" | "projects">("root");
     const [open, setOpen] = useState<boolean>(false);
     const [search, setSearch] = useState("");
 
@@ -28,7 +27,7 @@ const Example = () => {
                         children: "Projects",
                         closeOnSelect: false,
                         onClick: () => {
-                            setPage("projects");
+                            alert("projects");
                         },
                     },
                 ],
@@ -67,37 +66,29 @@ const Example = () => {
             <button onClick={() => setOpen(true)}>
                 Search
             </button>
-            <CommandPalette
+            <Search
                 onChangeSearch={setSearch}
                 onChangeOpen={setOpen}
                 search={search}
                 isOpen={open}
-                page={page}
             >
-                <CommandPalette.Page id="root">
-                    {filteredItems.length ? (
-                        filteredItems.map((list) => (
-                            <CommandPalette.List key={list.id} heading={list.heading}>
-                                {list.items.map(({ id, ...rest }) => (
-                                    <CommandPalette.ListItem
-                                        // className="bg-gohan"
-                                        key={id}
-                                        index={getItemIndex(filteredItems, id)}
-                                        {...rest}
-                                    />
-                                ))}
-                            </CommandPalette.List>
-                        ))
-                    ) : (
-                        <CommandPalette.FreeSearchAction />
-                    )}
-                </CommandPalette.Page>
+                {filteredItems.length ? (
+                    filteredItems.map((list) => (
+                        <Search.List key={list.id} heading={list.heading}>
+                            {list.items.map(({ id, ...rest }) => (
+                                <Search.ListItem
+                                    key={id}
+                                    index={getItemIndex(filteredItems, id)}
+                                    {...rest}
+                                />
+                            ))}
+                        </Search.List>
+                    ))
+                ) : (
+                    <Search.FreeSearchAction />
+                )}
 
-                <CommandPalette.Page id="projects">
-                    {/* Projects page */}
-                    Projects
-                </CommandPalette.Page>
-            </CommandPalette>
+            </Search>
         </>
     );
 };
